Add NbtLongArray and chunk heightmap calculation

+++ src/common/Redstone.NBT/NbtBinaryReader.cs
@@ -33,7 +33,7 @@ namespace Redstone.NBT
             {
                 throw new EndOfStreamException();
             }
-            else if (type > (int)NbtTagType.IntArray)
+            else if (type > (int)NbtTagType.LongArray)
             {
                 throw new NbtFormatException("NBT tag type out of range: " + type);
             }
